requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
reportlab==4.2.5
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
